+++ Code/Basic/Linux/Memory.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void *AllocatePlatformMemory(size_t size);
-void FreePlatformMemory(void *memory, size_t size);
+void *AllocatePlatformMemory(s64 size);
+void FreePlatformMemory(void *memory, s64 size);
 size_t GetPageSize();
