+++ Code/Basic/Linux/Semaphore.cpp
@@ -1,4 +1,4 @@
-Semaphore CreateSemaphore(u32 initialValue)
+Semaphore CreateSemaphore(s64 initialValue)
 {
 	sem_t semaphore;
 	sem_init(&semaphore, 0, initialValue);
@@ -15,7 +15,7 @@ void WaitOnSemaphore(Semaphore *semaphore)
 	sem_wait(semaphore);
 }
 
-s32 GetSemaphoreValue(Semaphore *semaphore)
+s64 GetSemaphoreValue(Semaphore *semaphore)
 {
 	s32 value;
 	sem_getvalue(semaphore, &value);
