+++ Code/Basic/Linux/Semaphore.h
@@ -4,8 +4,7 @@
 
 typedef sem_t Semaphore;
 
-Semaphore CreateSemaphore(u32 initialValue);
+Semaphore CreateSemaphore(s64 initialValue);
 void SignalSemaphore(Semaphore *semaphore);
 void WaitOnSemaphore(Semaphore *semaphore);
-s32 GetSemaphoreValue(Semaphore *semaphore);
-
+s64 GetSemaphoreValue(Semaphore *semaphore);
